Fix autoport issue
DEVSIX-1195

+++ layout/src/main/java/com/itextpdf/layout/renderer/AbstractRenderer.java
@@ -1022,7 +1022,7 @@ public abstract class AbstractRenderer implements IRenderer {
             width -= calculatePaddingBorderWidth(this);
         }
 
-        return width != null ? Math.max(0, (float) width) : null;
+        return width != null ? (Float) Math.max(0, (float) width) : null;
     }
 
     /**
